WIP: Add layer metrics
This is an early prototype to explore the Keras `layers.add_metrics` API for our stateful metrics as a replacements for the `QuantizedLogger`. I think this approach is a lot more flexible and elegant than using a callback.

Before I'll clean this PR up, we should check if this approach has a big performance impact since it needs to keep track of the previous quantized weights.