Move larq_zoo docs to main repo
Adding documentation for a new larq zoo model can be a bit cumbersome because its API docs have to be manually added and there is no good way to preview the changes to the website.
This moves the larq_zoo docs into the main repo to allow for easier developement. Also API docs are now automatically generated using the same process as larq it self.
While this might