Add JOSS publication
This is a first draft of our submission to [The Journal of Open Source Software](https://joss.theoj.org). The aim is for a paper that gives a quick overview of the software package within 250-1000 words.
I'd be very happy to get feedback into which direction we should go with this paper and where we need to evaluate.