General clarification of Key concepts page (#254)
* General clarification of Key concepts

Reading the documentation should follow the cognitive flow of the user. 
There were several issues. Examples: Quantization was not defined and directly after used in multiple senses (the concept and the `larq` implementation), which are not identical and the last section refers to QNN, without naming it first.

Some of the proposed changes are style preferences. Feel free to use any of the suggestions.

* Update docs/guides/key-concepts.md

Co-Authored-By: Koen Helwegen <kg.helwegen@gmail.com>

* Update docs/guides/key-concepts.md

Co-Authored-By: Koen Helwegen <kg.helwegen@gmail.com>