User integer as kernel_quantizer to mark layers that should use Bop
As discussed in https://github.com/larq/larq/issues/306#issuecomment-553978646, this allows to use integers as `kernel_quantizers` in layers to mark them to be used by external optimizers like Bop.

This will be a breaking change because previously `Bop` was applied to all quantized layers which made it impossible to use `larq` layers with different precisions in a network. Now `kernel_quantizers=1` needs to be set on all layers that should be optimized by `Bop`. `kernel_quantizers="ste_sign"` keeps working too, but will raise a warning message.

Closes #306

Co-Authored-By: Leon Overweel <l.overweel@gmail.com>